added tests

+++ migrations/sqls/20230117140633-order-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE order_products (
     id SERIAL PRIMARY KEY,
     quantity INTEGER,
-    order_id BIGINT REFERENCES orders(id),
-    product_id BIGINT REFERENCES products(id)
+    order_id INTEGER REFERENCES orders(id),
+    product_id INTEGER REFERENCES products(id)
 );
